# performance fix: String join