The code generated here for Stateful is (necessarily) quite slow (and expensive to compile)
and not really even any shorter or simpler to write. For reference, the expected performance report on @nanosoldier `runbenchmarks("string", vs = ":master")` is 20-40% (and afaict, is also going to be slightly faster than v0.6, but possibly below the noise threshold for reporting it):

```
julia> x = ("a", 2); @btime sprint(join, $x, "")
  1.301 μs (15 allocations: 560 bytes)
"a2"

julia> x = ("a", 2); @btime sprint(join2, $x, "")
  484.754 ns (8 allocations: 336 bytes)
"a2"

julia> x = ("a", "2"); @btime sprint(join, $x, "")
  1.396 μs (13 allocations: 464 bytes)
"a2"

julia> x = ("a", "2"); @btime sprint(join2, $x, "")
  364.522 ns (6 allocations: 240 bytes)
"a2"
```