@vtjnash 2018-07-06
> is 20-40% 

Actually, I'm not certain what that benchmark contains. If `x::Vector{String}`, then the improvement is only expected to be 75% (below reporting threshold of 60%). While if it's `x::Vector{Any}`, the improvement is expected to be 15%! (all relative to master, vs. 0.6 all of these seem pretty consistent to be about 75%, but that might be unrelated measurement error since there's so many other changes).