@nanosoldier 2018-07-06
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27974) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/6210049591eccd7ddf909e7dc5d73a6667a90625/e788ee0_vs_a60119b/report.md). cc @ararslan