@KristofferC 2018-07-07
@nanosoldier `runbenchmarks("problem", vs = ":master"),`