@nanosoldier 2018-07-07
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27974#issuecomment-403191497) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/f864bf224fc7105dd025ca8e582ae6cf3b0fdca3/825cb51_vs_6014e7d/report.md). cc @ararslan